feat: src: clean code
